+++ R/fill_yaml.R
@@ -1,4 +1,15 @@
-fill_yaml <- function(yml,
+fill_yaml <- function(## function-level args
+                      yml,
+                      verbose=TRUE,
+                      omit_defaults = c("tinytex_installer",
+                                        "tinytex_version",
+                                        "pandoc_version",
+                                        "miniforge_variant",
+                                        "miniforge_version",
+                                        "activate_environment",
+                                        "environment_file",
+                                        "channels"),
+                      enable_act,
                       ## action-level args
                       name,
                       template,
@@ -24,8 +35,12 @@ fill_yaml <- function(yml,
                       docker_user,
                       docker_org,
                       docker_token,
-                      cache_version,
-                      enable_act){
+                      cache_version, 
+                      miniforge_variant,
+                      miniforge_version,
+                      activate_environment,
+                      environment_file,
+                      channels){
   # devoptera::args2vars(use_workflow)
   
   #### name ####
@@ -51,16 +66,30 @@ fill_yaml <- function(yml,
       yml$jobs[[1]]$steps[[1]] <- NULL
     }  
   }
+  #### Check Miniconda args ####
+  miniconda_args <- check_miniconda_args( 
+    runners=runners,
+    miniforge_variant=miniforge_variant,
+    miniforge_version=miniforge_version,
+    activate_environment=activate_environment,
+    environment_file=environment_file,
+    channels=channels,
+    verbose=verbose)
+  miniforge_variant <- miniconda_args$miniforge_variant
+  miniforge_version <- miniconda_args$miniforge_version
+  activate_environment <- miniconda_args$activate_environment
+  environment_file <- miniconda_args$environment_file
+  channels <- miniconda_args$channels  
   #### Supply variables as "with:" or "env:" ####
-  nonarg_list <- c("yml","name","template","tag","on","branches","runners")
+  nonarg_list <- c("yml","verbose","omit_defaults","enable_act",
+                   "name","template","tag","on","branches","runners")
   args_list <- names(formals(fill_yaml))
-  args_list <- args_list[!args_list %in% nonarg_list]
-  omit_list <- c("tinytex_installer","tinytex_version","pandoc_version")
+  args_list <- args_list[!args_list %in% nonarg_list] 
   #### Omit certain variables when equal to default ####
   ## Don't do this for rworkflows_static as this setup has no default values.
   for(a in args_list){
     nm <- if(a %in% c("github_token","docker_token")) toupper(a) else a
-    with2[nm] <- if(a %in% omit_list &&
+    with2[nm] <- if(a %in% omit_defaults &&
                     template!="rworkflows_static") {
       omit_if_default(arg = a)
     } else {