+++ R/preview_yaml.R
@@ -1,0 +1,3 @@
+preview_yaml <- function(yml){
+  cat(yaml::as.yaml(yml))
+}