+++ R/return_yaml.R
@@ -1,0 +1,19 @@
+return_yaml <- function(yml,
+                        path,
+                        return_path,
+                        verbose=TRUE){
+  if(!is.null(path)){  
+    save_yaml(yml=yml,
+              path=path,
+              verbose=verbose)
+    #### Return ####
+    if(isTRUE(return_path)){
+      return(path)
+    } else {
+      yml <- yaml::read_yaml(path)
+      return(yml)
+    }
+  } else {
+    return(yml)
+  }
+}